@@ -1,6 +1,6 @@
 name: firestorehelpertest
 description: A new Flutter project.
-
+publish_to: none
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
 # followed by an optional build number separated by a +.
@@ -10,13 +10,12 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
  
 dependencies:
   cloud_firestore: any
-  jaguar_serializer: any 
   firestore_helpers:
-    path: ..\
+    path: ../
   flutter:
     sdk: flutter
 
