@@ -1,6 +1,6 @@
 name: firestore_helpers
 description: Firestore helper function to create dynamic and location based queries
-version: 3.2.2
+version: 4.0.0
 
 maintainer: Thomas Burkhart (@escamoteur)
 authors:
@@ -9,13 +9,12 @@ authors:
 homepage: https://github.com/fluttercommunity/firestore_helpers
 documentation: null
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  cloud_firestore: ^0.14.2
-  meta: ^1.2.3
+  cloud_firestore: ^1.0.1
 
 dev_dependencies:
   flutter_test:
