@@ -13,11 +13,11 @@ class QueryConstraint {
   final dynamic isLessThanOrEqualTo;
   final dynamic isGreaterThan;
   final dynamic isGreaterThanOrEqualTo;
-  final bool isNull;
+  final bool? isNull;
   final dynamic arrayContains;
 
   QueryConstraint(
-      {this.field,
+      {required this.field,
       this.isEqualTo,
       this.isLessThan,
       this.isLessThanOrEqualTo,
@@ -44,10 +44,7 @@ class OrderConstraint {
 /// [constraints] : a list of constraints that should be applied to the [collection].
 /// [orderBy] : a list of order constraints that should be applied to the [collection] after the filtering by [constraints] was done.
 /// Important all limitation of FireStore apply for this method two on how you can query fields in collections and order them.
-Query buildQuery(
-    {Query collection,
-    List<QueryConstraint> constraints,
-    List<OrderConstraint> orderBy}) {
+Query buildQuery({required Query collection, List<QueryConstraint>? constraints, List<OrderConstraint>? orderBy}) {
   Query ref = collection;
 
   if (constraints != null) {
@@ -70,7 +67,7 @@ Query buildQuery(
   return ref;
 }
 
-typedef DocumentMapper<T> = T Function(DocumentSnapshot document);
+typedef DocumentMapper<T> = T? Function(DocumentSnapshot document);
 typedef ItemFilter<T> = bool Function(T);
 typedef ItemComparer<T> = int Function(T item1, T item2);
 
@@ -84,14 +81,14 @@ typedef ItemComparer<T> = int Function(T item1, T item2);
 /// on the result on the client side
 /// [orderComparer] : optional comparisson function. If provided your resulting data
 /// will be sorted based on it on the client
-Stream<List<T>> getDataFromQuery<T>({
-  Query query,
-  DocumentMapper<T> mapper,
-  List<ItemFilter<T>> clientSidefilters,
-  ItemComparer<T> orderComparer,
+Stream<List<T >> getDataFromQuery<T>({
+  required Query query,
+  required DocumentMapper<T > mapper,
+  List<ItemFilter<T>>? clientSidefilters,
+  ItemComparer<T>? orderComparer,
 }) {
   return query.snapshots().map((snapShot) {
-    Iterable<T> items = snapShot.docs.map(mapper);
+    Iterable<T> items = snapShot.docs.map(mapper).where((element) => element != null) as Iterable<T>;
     if (clientSidefilters != null) {
       for (var filter in clientSidefilters) {
         items = items.where(filter);
